Bug#21624851: Assertion failed: is_prepared && !is_optimized()
This problem occurs on execution of a prepared statement.
The root cause is that a call to THD::replace_rollback_place_for_ref()
in SELECT_LEX::simplify_joins() records an invalid rollback "place"
for an item representing a predicate: An AND node is built from
an old AND node and a join condition. A rollback place is recorded
for one of the conditions of the old AND node. But when the rollback
is carried out, the "place" is pointing to the old AND node, which is
then replaced by an old Item_field pointer. What should happen is that
when the predicate is moved to the new AND node, we should track the
"place" inside the new AND node instead of the old one.
And this is what THD::replace_rollback_place_for_value() does.

The solution is thus to call THD::replace_rollback_place_for_ref() for
all arguments of the AND node.

It was also detected a wrong call to THD::change_item_tree() within
check option processing, which could be replaced with simple
assignment.

Test cases for several duplicate bug reports have been added:
21624851, 21626495, 21624331, 21626722, 21626177.