Testcase prep : Add TransporterRegistry blockSend test functionality
The TransporterRegistry already supports blocking
receive from a particular transporter,
blocking the reception of signals and e.g.
socket closures.

This patch adds blocking of send, so that sends
will be buffered by SendBuffers, but not sent over
the socket.  Note that this blocking does not
currently affect socket closure handling, or signal
reception.