Bug#21762928 mysys_my_symlink-t fails on Mac
The test fails on OS X because it fails to consider that a filename
handed back by create_temp_file may be based on a symbolic link. In
this case, /var -> /private/var.

Adding an extra call to my_realpath before conparing with the resolved
symbolic link created by the test itself solved the problem.