Bug#21067109: Assert 'join == __NULL' failed in ::optimize()
Patch # 1 of 5.

This is an issue with missing propagation of error values.
A const subquery is evaluated when removing const items in
remove_eq_conds(). The subquery fails because it returns more than
one row. This error code is poorly handled and causes the optimizer
to attempt to optimize a subquery a second time, which causes
this assert.

This patch changes interfaces for eval_const_cond(), remove_eq_conds()
and internal_remove_eq_conds(). The interfaces encourage better error
propagation but requires slightly more code to be written.