Bug#21067109: Assert 'join == __NULL' failed in ::optimize()
Patch # 2 of 5.

This patch refactors the interfaces for optimize_cond() and
related functions. All functions get a THD argument and a bool return
value that indicates success or error. These functions are refactored:

- optimize_cond()
- check_simple_equality()
- check_row_equality()
- check_equality()
- build_equal_items()
- build_equal_items_for_cond()
- change_cond_ref_to_const()
- propagate_cond_constants()

In addition, these changes are done:

- All errors are processed correctly
- C casts are replaced with down_cast<>

In addition, the logic of optimize_cond() is slightly changed:

- For HAVING clause, it is only called when HAVING != NULL.
- For WHERE clause, it is only called when WHERE != NULL or
  there is a least one outer join (this is called after simplify_joins()
  so if we have join nests, we know at least one join condition exists).

This is a preparation for a small optimization, which will come in patch#4.

Item_hex_string::clone_item() was implemented.