BUG#21755699: GR+XA: ASSERT `M_TRANSACTION_CTX.M_SIDNO == 0 && M_TRANSACTION_CTX.M_GNO == 0'
XA transactions were hitting a assert condition on XA COMMIT, this
was happening because despite XA transaction is composed by two
transactions: XA PREPARE; XA COMMIT/ROLLBACK; for engines and binary
log. The internal transaction state is not reset between PREPARE and
COMMIT/ROLLBACK, what was causing the assert on Rpl_transaction_ctx
to be triggered.
A second issue was found, XA ROLLBACK was not being properly
intercepted.

This is now fixed by removing the assert, since we may have two
consecutive calls of method
Rpl_transaction_ctx::set_rpl_transaction_ctx() for the same
transaction when it is XA.
Also XA ROLLBACK is properly intercepted on this patch.