Improve handling of class loader for clusterj
ClusterJHelper uses the Thread.getContextClassLoader() to obtain a
class loader in order to load the SessionFactory. SessionFactory then uses
Thread.getContextClassLoader() to obtain a class loader
in order to load the ClusterConnection. This strategy is flawed and is
now improved.

The new strategy:

ClusterJHelper uses a supplied class loader to load the SessionFactory.
If none is supplied, ClusterJHelper uses its own class loader.
SessionFactory uses its own class loader to load the ClusterConnection.

The improved strategy allows for the user to specify the class loader
to be used for the SessionFactory and ClusterConnection. If not specified,
the ClusterJHelper class loader is used to load all classes and resources.

ClusterJHelper.java
  Add method getServiceInstance that takes a service class and implementation
    class name and uses ClusterJHelper class loader as a default
  Replace all occurrences of thread context class loader with own class loader

SessionFactoryImpl.java
  Use new getServiceInstance method in ClusterJHelper to get implementation
    class for ClusterConnection.

DomainTypeHandlerFactoryImpl.java
  Use own class loader to get implementation classses for DomainTypeHandlers.

AbstractClusterJCoreTest.java
  Use own class loader instead of thread context class loader to load resources

AbstractClusterJTest.java
  Use own class loader instead of thread context class loader to load resources