Bug#21774859 CMAKE CHECKS FOR NUMA SUPPORT TOO WEAK
The check for numa availability implemented by the patch for
Bug #18871046 SET NUMA MEMPOLICY FOR OPTIMUM MYSQLD PERFORMANCE
was wrong: it forgot to restore CMAKE_REQUIRED_LIBRARIES after the check.

This affects all subsequent platform feature tests that rely on cmakes
CHECK_CXX_SOURCE_COMPILES. In particular: on platforms which do not have
numa support, the test for 'HAVE_PEERCRED' will also fail, and the
'auth_socket' plugin will not be built.