Bug#11758430: NDB_SHOW_TABLES SHOULD REQUIRE A DATABASE WHEN TABLE IS SUPPLIED
When the ndb_show_tables command was run without a database
explicitly specified, it searched TEST_DB by default. A failure
message was displayed if the table was not present. The error
message did not specify that this was different from the normal
usage of the ndb_show_tables command. It returned the output of
ndb_show_tables command if the table did in fact exist in TEST_DB,
otherwise a message informing the user that the specified table did
not exist was displayed.

Modified the behaviour of ndb_show_tables to track when TEST_DB is
being checked in the absence of any specified database. Now, it
indicates that a database has to be specified along with the command
in case the table is not present in TEST_DB.