Bug #21480999: FAILING ASSERTION: LEN <= FIXED_LEN IN                REC_GET_CONVERTED_SIZE_COMP_PREFIX_LOW
Issue:
-----
When creating a temporary table, the length of the largest
column in the union is taken into consideration. In a union
operation on a fixed length column like CHAR and a SET
column, the largest column's byte-length can exceed 1023.
This creates a problem when the temporary table is created
in Innodb, since fixed length columns are supposed to be
smaller than 1024.

SOLUTION:
---------
If the number of bytes in the column can exceed the product
of the maximum number of characters and the charset
encoding's byte-length, use a variable length column. This
will avoid any issue with the temporary table created in
Innodb.