Bug#21215389: UPGRADE GOOGLE MOCK TO 1.7
This patch changes the server to use Google Mock 1.7 rather
than 1.6 for unit testing. The patch also removes now obsolete
workarounds for 1.6 specific compilation issues.

The patch also updates the behavior of the WITH_GMOCK CMake option.
If WITH_GMOCK points to a .zip, that .zip is used if it matches
the required Google Mock version (now 1.7). If a .zip is found,
source_downloads in the source dir is used for unpacking.

If WITH_GMOCK points to a directory, that directory is searched
for Google Mock versions (both .zip and unpacked). If no versions
are found, that directory is used for downloading
(if ENABLE_DOWNLOADS is used). Afterwards, the same directory
is used for unpacking of Google Mock. Note that releative paths
are not supported for WITH_GMOCK.