Bug#21689998 ST_UNION() RETURNS AN INVALID GEOMETRYCOLLECTION
The bug is due to the way Boost.Geometry handles linear geometries that have
spikes when performing the difference and intersection set operations with
areal geometries. By default Boost.Geometry discards spikes present in linear
geometries, which can easily result to set operation results that are invalid
(linestring with topological dimension 0 instead of 1).

Fix: deactivate the removal of spikes present in linear geometries when performing
the difference(L,A) and intersection(L,A) set operations.

Note: The Boost.Geometry related patch that is part of this commit has been submitted for
inclusion to Boost.Geometry; it has not been reviewed yet, so it may change as part of the
Boost.Geometry reviewing process.