Bug#21658453 ST_INTERSECTION() PRODUCES INVALID POLYGON
The root cause of the bug is that the Boost.Geometry intersection algorithm does
not take into account certain intersection points of the geometries to be
intersected; as a result, in some cases, the rings produced by the Boost.Geometry
intersection algorithm contain self-intersection points and do not have the correct
topology.

Fix: add a post-processing step to the Boost.Geometry intersection algorithm which
takes care of these self intersections; as part of the post-processing step
points may be added on the rings already computed and ring may be split into
smaller rings that do not have self intersection points.

The Boost.Geometry difference and sym_difference algorithms suffer from the same
problem; the Boost.Geometry patch committed as part of this commit takes care of
these problems as well.

New test cases related to the bug report have been added, and previous results have
been updated. As far as previous results are concerned, it is interesting to note
that the previous results were invalid geometries, whereas the newly recorded ones
correspond to valid geometries (as expected and wanted).

The patch pushed as part of this commit has not yet been pushed into the
Boost.Geometry develop branch and it is still under review. As such, it may
undergo changes, in which case any updates will be migrated to MySQL.