BUG#21780891 TRANSACTION CONSUMERS ENABLED BUT NOT THE INSTRUMENT
Befire this fix,

By default, the transaction instrument is disabled in table
setup_instruments.

By default, the transactions consumers "transaction_current"
and "transaction_history" are enabled in table setup_consumers.

The net effect is that transactions are not instrumented in any case,
because the instrument is disabled, so displaying "enabled"
consumers by default is misleading.

The fix is to disable all transactions related consumers by default.

This is an ease of use fix, which has otherwise no impact.