Bug#21205282 CRASH/ASSERTION IN JOIN_CACHE::SET_MATCH_FLAG_IF_NONE WITH NO_BNL HINT
There is limitation of JOIN_CACHE in case when
table belongs to nested outer join. If an inner
table of an outer join does not use JOIN_CACHE,
all other inner tables of that join must not use
JOIN_CACHE either.

The fix is to use revise_cache_usage() to enforce the rule above,
if a hint prevents join buffering for one table.