BUG#21680074 - SOME COMMAND SHOULD NOT BE ALLOWED FOR CHANNEL 'GROUP_REPLICATION_RECOVERY'
There is a set of commands that should not be allowed for
group_replication_recovery channel as 'group_replication_recovery' is is
not exactly a slave channel.

Looking quickly to the available commands and pursuing a mistake free
environment, we should disable the following commands:

START/STOP SLAVE
START/STOP IO_THREAD|SQL_THREAD

The following commands are blocked from being used when the channel is
'group_replication_recovery'.