BUG#21798804: GR ALLOWS UNSUPPORTED MTS TYPE 'DATABASE' POST SQL_THREAD RESTART
Group Replication (GR) applier channel does not support
SLAVE_PARALLEL_TYPE set to DATABASE, when GR is started that is
tested and a error is thrown. Though SLAVE_PARALLEL_TYPE can be
changed with the following steps:
  STOP SLAVE SQL_THREAD FOR CHANNEL 'group_replication_applier';
  SET GLOBAL SLAVE_PARALLEL_WORKERS=4;
  START SLAVE SQL_THREAD FOR CHANNEL 'group_replication_applier';
which is incorrect and will cause applier to error out.

This patch fixes the above issue by always checking that
SLAVE_PARALLEL_TYPE on group_replication_applier is LOGICAL_CLOCK
when the number of SLAVE_PARALLEL_WORKERS is greater then 0.