Bug#21574933 Extra rows with derived table in subquery + XOR
This problem stems from the transformation of IN subqueries
in Item_in_subselect::single_value_in_to_exists_transformer().

The source of the regression is the fix for bug no. 14358878, which
wrapped all selected columns from a derived table that was on the
inner side of an outer join in Item_direct_view_ref objects.
The purpose of the wrapping was to make the columns nullable and
nullability depending on some table from the derived table.

single_value_in_to_exists_transformer() calculated orig_item from one
such column. Prior to the wrapping, nullability information was present
in orig_item. But after the bugfix, real_item() would pass the
Item_direct_view_ref object and return the underlying Item object,
which is not necessarily nullable.

real_item() was originally used so that we do not build a permanent
subquery transformation on top of a runtime item. The wrapped
Item_direct_view_ref objects are runtime objects, meaning that
they cannot be used in a permanent transformation.
The solution is a kludge: If the Item_direct_view_ref object is
nullable, we set the item being pointed to from the
Item_direct_view_ref as nullable as well. This means that we pick up
correct nullability, but the item does not correctly reflect the
nullability of the datum. However, we have not been able to identify
a situation where this causes erroneous or non-optimal execution.

Further on, this means that NULL values from outer joined tables
are included in the query result, meaning that the IN subquery
can return NULL instead of FALSE in cases like this.