Bug#21779011 INVALID READS AND SENDING RANDOM SERVER MEMORY BACK TO CLIENT
When a table containing a virtual generated column was stored in the
InnoDB partition engine, reading from this table could in some cases
return random data for the virtual generated column. The cause for
this, was that the code that calculated and filled the value for the
virtual generated column was not run.

The rows are read using handler::multi_range_read_next() which calls
read_range_first(). handler::read_range_first() is a virtual
function. In the case of most other storage engines, the version
implemented in the handler class is used. This function calls
handler::ha_index_read_map() which updates the generated column
values. When using the InnoDB partition engine, the specialized
version implemented in ha_innopart::read_range_first() is used. This
function does not use handler::ha_index_read_map() for reading the
index but instead uses corresponding functions implemented in
ha_innobase. None of these functions update the generated column
values. The result is that when returning from
handler::multi_range_read_next(), the virtual generated columns have
not been updated.

The fix for this is to add updating of virtual generated columns in
handler::multi_range_read_next(). To avoid that the virtual generated
columns are updated multiple times (eg both in
handler::multi_range_read_next() and handler::ha_index_read_map()) a
new flag is added to the handler class that keeps track of whether the
generated columns have been updated or not.